backend/pyproject.toml
[project]
name = "recipe-agent-backend"
version = "0.1.0"
description = "Microsoft Agent Framework backend with AG-UI protocol for interactive recipes"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.124.2",
    "uvicorn==0.38.0",
    "python-dotenv>=1.2.1",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "agent-framework==1.0.0b251209",
    "agent-framework-ag-ui==1.0.0b251209",
    "sse-starlette>=3.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
